2b: Split subagent.go into subagent.go + subagent_drain.go
